src/controllers/interrupt.card.controller.ts
import { NextFunction, Request, Response } from "express";

import { logger } from "../utils/logger";
import * as config from "../config";
import { deleteApplicationData } from "../utils/application.data";

export const get = (req: Request, res: Response, next: NextFunction) => {
  try {
    logger.debugRequest(req, `GET ${config.INTERRUPT_CARD_PAGE}`);

    deleteApplicationData(req.session);
    return res.render(config.INTERRUPT_CARD_PAGE);
  } catch (error) {
    logger.errorRequest(req, error);
    next(error);
  }
};

export const post = (req: Request, res: Response, next: NextFunction) => {
  try {
    logger.debugRequest(req, `POST ${config.INTERRUPT_CARD_PAGE}`);
    return res.redirect(config.PRESENTER_URL);
  } catch (error) {
    logger.errorRequest(req, error);
    next(error);
  }
};
